+++ src/calibration.hpp
@@ -13,11 +13,11 @@ public:
 
     // Run the non-linear frequency sweep and return optimal frequency.
     // Also captures optimal physics parameters if desired.
-    bool run_sweep(long double& out_omega);
+    bool run_sweep(double& out_omega);
 
     // Persistence Layer
-    static bool save_profile(const std::string& path, long double omega, long double alpha, long double beta, long double delta);
-    static bool load_profile(const std::string& path, long double& omega, long double& alpha, long double& beta, long double& delta);
+    static bool save_profile(const std::string& path, double omega, double alpha, double beta, double delta);
+    static bool load_profile(const std::string& path, double& omega, double& alpha, double& beta, double& delta);
 };
 
 } // namespace core
